Reading line and char count from nodes

+++ git_analysis/GraphNode.h
@@ -27,7 +27,7 @@ namespace GB {
 		}
 
 		void is_root() { is_root_ = true; }
-		void get_line_and_char_count(uint32_t &line_count, uint32_t &char_count);
+		void get_line_and_char_count(uint64_t &line_count, uint64_t &char_count);
 		bool process();
 	private:
 		std::shared_ptr<GraphNode> parent_{};
@@ -35,8 +35,8 @@ namespace GB {
 		std::map<std::shared_ptr<GraphNode>, uint32_t> child_map_{};
 		std::string name_{};
 		bool is_root_{ false };
-		uint32_t line_count_{};
-		uint32_t character_count_{};
+		uint64_t line_count_{};
+		uint64_t character_count_{};
 		bool read_file_{ false };
 	};
 }
